Add files via upload
There is 3 parts to this project . It is given the propability , assortment and assortment test txt. Our analysis is the Part a, b ,c python file and our output for part a is group6.txt 
More information about the final project questions in the final project pdf file . 

Part (a) requires predicting the choice probability for products in given assortments, with predictions recorded in a specific format. Part (b) focuses on determining the optimal price for product 'b' when it's the only item carried, with the flexibility to propose three different pricing strategies. Finally, part (c) challenges the identification of the most revenue-optimizing assortment of up to three products, given specific prices for six SKUs, with the allowance to submit three different assortment solutions. The overall goal is to utilize data-driven insights to enhance decision-making in product selection and pricing.